move to gammapy.utils.roots
