+++ gammapy/utils/gauss.py
@@ -2,6 +2,7 @@
 """Multi-Gaussian distribution utilities (Gammapy internal)."""
 import numpy as np
 from astropy import units as u
+from gammapy.utils.roots import find_roots
 
 class Gauss2DPDF:
     """2D symmetric Gaussian PDF.
@@ -262,7 +263,6 @@ class MultiGauss2D:
         containment_radius : `~numpy.ndarray`
             Containment radius
         """
-        from gammapy.estimators.utils import find_roots
         rad_max = 1e3
 
         def f(rad):
