+++ gammapy/utils/scripts.py
@@ -148,3 +148,5 @@ def recursive_merge_dicts(a, b):
         else:
             c[k] = v
     return c
+
+
